requirements.txt
six==1.10.0
pyzmq==16.0.2
pytest==3.0.6

